requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.3
idna==2.6
kombu==4.1.0
pkg-resources==0.0.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
urllib3==1.22
vine==1.1.4
